requirements.txt
django==5.2.*
environs[django]==14.2.*
django-phonenumber-field==8.1.*
phonenumbers==9.0.*
pillow==11.3.*
gunicorn==23.0.*
psycopg2-binary==2.9.*